% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boots.R
\name{boot}
\alias{boot}
\alias{boot.glm_cal}
\alias{boot.gam_cal}
\alias{boot.lowess_cal}
\alias{boot.loess_cal}
\title{Bootstrap resample a calibration curve object}
\usage{
boot(cal)

\method{boot}{glm_cal}(cal)

\method{boot}{gam_cal}(cal)

\method{boot}{lowess_cal}(cal)

\method{boot}{loess_cal}(cal)
}
\arguments{
\item{cal}{an object created using one of the \code{cal} functions}
}
\value{
bootstrap resamples of calibration metrics and values for plotting
}
\description{
Bootstrap resample a calibration curve object
}
\keyword{internal}
