% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_method_smooth}
\alias{check_method_smooth}
\title{Check if method and smooth are allowed together (not used anymore)}
\usage{
check_method_smooth(method, smooth)
}
\description{
Check if method and smooth are allowed together (not used anymore)
}
\keyword{internal}
