% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_cal.R
\name{glm_cal}
\alias{glm_cal}
\title{fits a calibration curve via glm}
\usage{
glm_cal(y, p, X, Xp, save_data = T, save_mod = T)
}
\arguments{
\item{y}{binary outcome}

\item{p}{predicted probabilities}

\item{X}{design matrix created via \code{reg_spline_X} (rows of \code{X} should align with elements of \code{p})}

\item{Xp}{design matrix for to-be-plotted probability}

\item{save_data}{whether to save the data elements in the returned object}

\item{save_mod}{whether to save the model in the returned object}
}
\value{
list of class \code{glm_cal}
}
\description{
fits a calibration curve via glm
}
\keyword{internal}
