% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_cal.R
\name{logistic_cal}
\alias{logistic_cal}
\title{Run logistic calibration}
\usage{
logistic_cal(y, p)
}
\arguments{
\item{y}{binary outcome}

\item{p}{predicted probabilities (these will be logit transformed)}
}
\value{
an object of class \code{logistic_cal} containing \code{glm} results for calculating calibration intercept and calibration slope
}
\description{
Assess 'weak' calibration (see, e.g., Van Calster et al. 2019) via calibration intercept
and calibration slope.
}
\references{
Van Calster, B., McLernon, D. J., Van Smeden, M., Wynants, L., & Steyerberg, E. W. (2019). Calibration: the Achilles heel of predictive analytics. BMC medicine, 17(1), 1-7.
}
