% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.pmcalibration}
\alias{plot.pmcalibration}
\title{Plot a calibration curve}
\usage{
\method{plot}{pmcalibration}(x, ...)
}
\arguments{
\item{x}{a \code{pmcalibration} calibration curve}

\item{...}{other args for plot()}
}
\description{
This is for a quick and dirty calibration curve plot.
Alternatively you can use \code{get_cc()} to get the data required to plot the calibration curve.
}
