% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcalibration-package.R
\docType{package}
\name{pmcalibration-package}
\alias{pmcalibration-package}
\alias{_PACKAGE}
\title{pmcalibration: Calibration Curves For Clinical Prediction Models}
\description{
Fit calibrations curves for clinical prediction models and calculate several associated metrics (Eavg, E50, E90, Emax). Ideally predicted probabilities from a prediction model should align with observed probabilities. Calibration curves relate predicted probabilities (or logit(probability)) to observed outcomes via a flexible non-linear smoothing function. `pmcalibration` allows users to choose between several smoothers (regression splines, GAMs, lowess, loess). Currently only binary outcomes are supported but we aim to expand functionality (e.g., include time-to-event).
}
\author{
\strong{Maintainer}: Stephen Rhodes \email{steverho89@gmail.com}

}
\keyword{internal}
