% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcalibration.R
\name{pmcalibration}
\alias{pmcalibration}
\title{Create a calibration curve}
\usage{
pmcalibration(
  y,
  p,
  smooth = c("none", "ns", "bs", "rcs", "gam", "lowess", "loess"),
  ci = c("sim", "boot"),
  conf_level = 0.95,
  n = 1000,
  logitp = T,
  neval = 100,
  ...
)
}
\arguments{
\item{y}{a binary outcome}

\item{p}{predicted probabilities from a clinical prediction model}

\item{smooth}{what smooth to use. For method = 'glm' and smooth = 'none' logistic calibration is performed}

\item{ci}{what kind of confidence intervals to compute? 'sim' = simulation based inference; \code{n} samples are taken from a multivariate normal distribution with mean vector = coef(mod) and variance covariance = vcov(model). 'boot' = bootstrap resampling with \code{n} replicates. \code{y} and \code{p} are sampled with replacement and calibration curve is reestimated. Calibration metrics are calculated using each simulation or boot sample. For both options percentile confidence intervals are returned.}

\item{conf_level}{width of the confidence interval (0.95 gives 95\% CI)}

\item{n}{number of simulations or bootstrap resamples}

\item{logitp}{fit calibration curve on logit transform of \code{p}}

\item{neval}{number of points (equally spaced between \code{min(p)} and \code{max(p)}) to evaluate for plotting (0 or NULL = no plotting)}

\item{...}{additional arguments for particular smooths. For ci = 'boot' the user is able to run samples in parallel (using the \code{parallel} package) by specifying a \code{cores} argument}

\item{method}{use regression splines (glm), generalized additive models (gam), or local regression smoothers (lowess, loess) to estimate calibration curve}
}
\value{
a \code{pmcalibration} object containing calibration metrics and values for plotting
}
\description{
Assess calibration of clinical prediction models (agreement between predicted and observed probabilities) via different smooths.
}
\references{
Austin PC, Steyerberg EW. (2019) The Integrated Calibration Index (ICI) and related metrics for quantifying the calibration of logistic regression models. \emph{Statistics in Medicine}. 38, pp. 1–15. https://doi.org/10.1002/sim.8281

Van Calster, B., Nieboer, D., Vergouwe, Y., De Cock, B., Pencina M., Steyerberg E.W. (2016). A calibration hierarchy for risk models was defined: from utopia to empirical data. \emph{Journal of Clinical Epidemiology}, 74, pp. 167-176
}
