% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowess_cal.R
\name{predict_lowess}
\alias{predict_lowess}
\title{Get predictions from \code{loewss} fit}
\usage{
predict_lowess(fit, x)
}
\arguments{
\item{fit}{list produced by \code{lowess}}

\item{x}{values to produce predictions for}
}
\value{
predicted values
}
\description{
Adapted from \code{rms:::calibrate.default}
Uses \code{approx} with \code{rule} = 2 so that x out of range of initial lowess fit returns min or max (see ?approx)
}
\keyword{internal}
