% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.logistic_cal}
\alias{print.logistic_cal}
\title{Print a \code{logistic_cal} object}
\usage{
\method{print}{logistic_cal}(x, digits = 2, conf_level = 0.95, ...)
}
\arguments{
\item{x}{a \code{logistic_cal} object}

\item{digits}{number of digits to print}

\item{conf_level}{width of the confidence interval (0.95 gives 95\% CI)}

\item{...}{optional arguments passed to print}
}
\value{
prints a summary
}
\description{
Print a \code{logistic_cal} object
}
