% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_cal.R
\name{print.logistic_calsummary}
\alias{print.logistic_calsummary}
\title{Print a logistic_cal summary}
\usage{
\method{print}{logistic_calsummary}(x, digits = 2, ...)
}
\arguments{
\item{x}{a \code{logistic_calsummary} object}

\item{digits}{number of digits to print}

\item{...}{ignored}
}
\value{
prints a summary
}
\description{
Print a logistic_cal summary
}
