% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.pmcalibrationsummary}
\alias{print.pmcalibrationsummary}
\title{Print summary of pmcalibration object}
\usage{
\method{print}{pmcalibrationsummary}(x, digits = 2, ...)
}
\arguments{
\item{x}{a \code{pmcalibrationsummary} object}

\item{digits}{number of digits to print}

\item{...}{ignored}
}
\value{
invisible(x) - prints a summary
}
\description{
Print summary of pmcalibration object
}
