% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boots.R
\name{run_boots}
\alias{run_boots}
\title{Wrapper to run bootstrap resamples using \code{parallel}}
\usage{
run_boots(cal, R = 1000, cores = 1)
}
\arguments{
\item{cal}{an object created with one of the \code{_cal} functions}

\item{R}{number of resamples (default = 1000)}

\item{cores}{number of cores (for \code{parallel})}
}
\value{
a list created by one of the \code{boot.} functions
}
\description{
Wrapper to run bootstrap resamples using \code{parallel}
}
\keyword{internal}
