% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simb.R
\name{simb}
\alias{simb}
\alias{simb.glm_cal}
\alias{simb.lowess_cal}
\alias{simb.loess_cal}
\alias{simb.gam_cal}
\title{Simulation based inference with a calibration curve object}
\usage{
simb(x, ...)

\method{simb}{glm_cal}(x, R = 1000)

\method{simb}{lowess_cal}(x, ...)

\method{simb}{loess_cal}(x, ...)

\method{simb}{gam_cal}(x, ...)
}
\arguments{
\item{x}{an object created using one of the \code{cal} functions}

\item{...}{other object specific arguments}
}
\value{
simulated calibration metrics and values for plotting
}
\description{
Simulation based inference with a calibration curve object
}
\keyword{internal}
