% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_cal.R
\name{gam_cal}
\alias{gam_cal}
\title{fits a calibration curve via gam}
\usage{
gam_cal(y, p, x, xp, save_data = T, save_mod = T, pw = F, ...)
}
\arguments{
\item{y}{binary outcome}

\item{p}{predicted probabilities}

\item{x}{predictor (could be transformation of \code{p})}

\item{xp}{values for plotting (same scale as \code{x})}

\item{save_data}{whether to save the data elements in the returned object}

\item{save_mod}{whether to save the model in the returned object}

\item{pw}{save pointwise standard errors for plotting}

\item{...}{additional arguments for \code{mgcv::gam} and \code{mgcv::s}}
}
\value{
list of class \code{gam_cal}
}
\description{
fits a calibration curve via gam
}
\keyword{internal}
