% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cc}
\alias{get_cc}
\title{Extract plot data from \code{pmcalibration} object}
\usage{
get_cc(x, conf_level = 0.95)
}
\arguments{
\item{x}{\code{pmcalibration} object}

\item{conf_level}{width of the confidence interval (0.95 gives 95\% CI). Ignored if call to \code{pmcalibration} didn't request confidence intervals}
}
\value{
data frame for plotting with 4 columns
\itemize{
\item{\code{p} - values for the x-axis (predicted probabilities - note these are *not* from your data and are only used for plotting)}
\item{\code{p_c} - probability impied by the calibration curve given \code{p}}
\item{\code{lower} and \code{upper} - bounds of the confidence interval
(type of CI and width determined by original call to \code{pmcalibration})}
}
}
\description{
Extract plot data from \code{pmcalibration} object
}
\examples{
\dontrun{
get_cc(cal) |>
 ggplot(aes(x = p, y = p_c, ymin=lower, ymax=upper)) +
 geom_abline(intercept = 0, slope = 1, lty=2) +
 geom_line() +
 geom_ribbon(alpha = 1/4)
 }
}
