% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.pmcalibration}
\alias{plot.pmcalibration}
\title{Plot a calibration curve}
\usage{
\method{plot}{pmcalibration}(x, conf_level = 0.95, ...)
}
\arguments{
\item{x}{a \code{pmcalibration} calibration curve}

\item{conf_level}{width of the confidence interval (0.95 gives 95\% CI). Ignored if call to \code{pmcalibration} didn't request confidence intervals}

\item{...}{other args for \code{plot()} (\code{lim} and \code{lab} can be specified)}
}
\description{
This is for a quick and dirty calibration curve plot.
Alternatively you can use \code{get_cc()} to get the data required to plot the calibration curve.
}
