% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcalibration-package.R
\docType{package}
\name{pmcalibration-package}
\alias{pmcalibration-package}
\title{pmcalibration: Calibration Curves for Clinical Prediction Models}
\description{
Fit calibrations curves for clinical prediction models and calculate several associated metrics (Eavg, E50, E90, Emax). Ideally predicted probabilities from a prediction model should align with observed probabilities. Calibration curves relate predicted probabilities (or a transformation thereof) to observed outcomes via a flexible non-linear smoothing function. 'pmcalibration' allows users to choose between several smoothers (regression splines, generalized additive models/GAMs, lowess, loess). Both binary and time-to-event outcomes are supported. See Van Calster et al. (2016) \doi{10.1016/j.jclinepi.2015.12.005}; Austin and Steyerberg (2019) \doi{10.1002/sim.8281}; Austin et al. (2020) \doi{10.1002/sim.8570}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stephenrho/pmcalibration}
  \item Report bugs at \url{https://github.com/stephenrho/pmcalibration/issues}
}

}
\author{
\strong{Maintainer}: Stephen Rhodes \email{steverho89@gmail.com} [copyright holder]

}
\keyword{internal}
