% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcalibration.R
\name{pmcalibration}
\alias{pmcalibration}
\title{Create a calibration curve}
\usage{
pmcalibration(
  y,
  p,
  smooth = c("none", "ns", "bs", "rcs", "gam", "lowess", "loess"),
  ci = c("sim", "boot", "pw", "none"),
  n = 1000,
  logitp = T,
  neval = 100,
  ...
)
}
\arguments{
\item{y}{a binary outcome}

\item{p}{predicted probabilities from a clinical prediction model}

\item{smooth}{what smooth to use. Available options:
\itemize{
\item{'rcs' = restricted cubic spline using \code{rms::rcs}. Optional arguments for this smooth are \code{nk} (number of knots; defaults to 5) and \code{knots} (knot positions; set by \code{Hmisc::rcs.eval} if not specified)}
\item{'ns' = natural spline using \code{splines::ns}. Optional arguments are \code{df} (default = 6), \code{knots}, \code{Boundary.knots} (see \code{?splines::ns})}
\item{'bs' = B-spline using \code{splines::bs}. Optional arguments are \code{df} (default = 6), \code{knots}, \code{Boundary.knots} (see \code{?splines::bs})}
\item{'gam' = generalized additive model via \code{mgcv::gam} and \code{mgcv::s}. Optional arguments are \code{bs}, \code{k}, \code{fx}, \code{method} (see \code{?mgcv::gam} and  \code{?mgcv::s})}
\item{'lowess' = uses \code{lowess(x, y, iter = 0)} based on \code{rms::calibrate}}
\item{'loess' = uses \code{loess} with all defaults}
\item{'none' = logistic regression with single predictor variable (performs logistic calibration when \code{logitp = T}). See \code{\link{logistic_cal}} }
}
'rcs', 'ns', 'bs', and 'none' are fit via \code{glm} and 'gam' is fit via \code{mgcv::gam} with \code{family = Binomial(link="logit")}}

\item{ci}{what kind of confidence intervals to compute?
\itemize{
\item{'sim' = simulation based inference; \code{n} samples are taken from a multivariate normal distribution with mean vector = coef(mod) and variance covariance = vcov(model).}
\item{'boot' = bootstrap resampling with \code{n} replicates. \code{y} and \code{p} are sampled with replacement and calibration curve is reestimated. If \code{knots} are specified the same knots are used for each resample (otherwise they are calculated using resampled x)}
\item{'pw' = pointwise confidence intervals calculated via the standard errors produced by relevant \code{predict} methods. Only for plotting curves - CIs are not produced for metrics (not available for smooth = 'lowess')}
}
Calibration metrics are calculated using each simulation or boot sample. For both options percentile confidence intervals are returned.}

\item{n}{number of simulations or bootstrap resamples}

\item{logitp}{fit calibration curve on logit transform of \code{p}}

\item{neval}{number of points (equally spaced between \code{min(p)} and \code{max(p)}) to evaluate for plotting (0 or NULL = no plotting)}

\item{...}{additional arguments for particular smooths. For ci = 'boot' the user is able to run samples in parallel (using the \code{parallel} package) by specifying a \code{cores} argument}
}
\value{
a \code{pmcalibration} object containing calibration metrics and values for plotting
}
\description{
Assess calibration of clinical prediction models (agreement between predicted and observed probabilities) via different smooths.
}
\references{
Austin PC, Steyerberg EW. (2019) The Integrated Calibration Index (ICI) and related metrics for quantifying the calibration of logistic regression models. \emph{Statistics in Medicine}. 38, pp. 1–15. https://doi.org/10.1002/sim.8281

Van Calster, B., Nieboer, D., Vergouwe, Y., De Cock, B., Pencina M., Steyerberg E.W. (2016). A calibration hierarchy for risk models was defined: from utopia to empirical data. \emph{Journal of Clinical Epidemiology}, 74, pp. 167-176
}
