% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simb.R
\name{simb}
\alias{simb}
\alias{simb.glm_cal}
\alias{simb.gam_cal}
\alias{simb.lowess_cal}
\alias{simb.loess_cal}
\title{Simulation based inference with a calibration curve object}
\usage{
simb(cal, R)

\method{simb}{glm_cal}(cal, R = 1000)

\method{simb}{gam_cal}(cal, R = 1000)

\method{simb}{lowess_cal}(cal, R)

\method{simb}{loess_cal}(cal, R)
}
\arguments{
\item{cal}{an object created using one of the \code{cal} functions}

\item{R}{number of simulated replicates}
}
\value{
simulated calibration metrics and values for plotting
}
\description{
Simulation based inference with a calibration curve object
}
\keyword{internal}
