% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.logistic_cal}
\alias{summary.logistic_cal}
\title{Summarize a logistic_cal object}
\usage{
\method{summary}{logistic_cal}(object, conf_level = 0.95, ...)
}
\arguments{
\item{object}{a \code{logistic_cal} object}

\item{conf_level}{width of the confidence interval (0.95 gives 95\% CI)}

\item{...}{ignored}
}
\value{
estimates and conf_level*100 confidence intervals for calibration intercept and calibration slope.
The former is estimated from a \code{glm} (family = binomial("logit")) where the linear predictor (logit(p)) is included as an offset.
}
\description{
Summarize a logistic_cal object
}
